refactor: pairing
